# storage: reduce allocations when populating BatchResponse