Instead of creating the response structs individually, we now loop over
the request and batch allocate all of the responses for a particular
type.

I'm not sure why there is so much variation in the allocations for the
Insert10000 benchmark.

```
name                  old time/op    new time/op    delta
KVInsert1_SQL-32         498µs ± 3%     504µs ± 2%  +1.24%  (p=0.001 n=20+20)
KVInsert10_SQL-32        971µs ± 2%     965µs ± 2%    ~     (p=0.175 n=19+20)
KVInsert100_SQL-32      5.34ms ± 4%    5.34ms ± 5%    ~     (p=0.840 n=19+19)
KVInsert1000_SQL-32     53.9ms ± 8%    53.9ms ±10%    ~     (p=0.779 n=20+20)
KVInsert10000_SQL-32     726ms ± 7%     705ms ±10%    ~     (p=0.091 n=17+20)

name                  old allocs/op  new allocs/op  delta
KVInsert1_SQL-32           316 ± 0%       314 ± 0%  -0.76%  (p=0.000 n=20+17)
KVInsert10_SQL-32          863 ± 0%       838 ± 0%  -2.81%  (p=0.000 n=20+20)
KVInsert100_SQL-32       6.15k ± 0%     5.94k ± 0%  -3.35%  (p=0.000 n=20+20)
KVInsert1000_SQL-32      59.5k ± 1%     57.5k ± 1%  -3.29%  (p=0.000 n=18+20)
KVInsert10000_SQL-32      690k ± 7%      684k ± 9%    ~     (p=0.305 n=19+20)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6222)
<!-- Reviewable:end -->